# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pinjie Fullstack Blog Backend",
    "description": "独立个人博客的后端 API，当前提供认证、账户、后台管理、文件资产和系统设置基础能力。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "后台管理"
        ],
        "summary": "获取用户列表",
        "operationId": "list_users_api_v1_admin_users_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "lifecycle",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "active",
                "inactive",
                "deleted"
              ],
              "type": "string",
              "default": "all",
              "title": "Lifecycle"
            }
          },
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_PageResult_AdminUserRead__"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "后台管理"
        ],
        "summary": "创建用户",
        "description": "由具备独立权限的管理员创建普通用户账户，不受公开注册开关影响且不创建登录会话。",
        "operationId": "create_user_api_v1_admin_users_post",
        "parameters": [
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserCreateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminUserRead_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/status/batch": {
      "patch": {
        "tags": [
          "后台管理"
        ],
        "summary": "批量修改用户状态",
        "operationId": "set_user_status_bulk_api_v1_admin_users_status_batch_patch",
        "parameters": [
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserBulkStatusUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_BatchActionResult_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/batch": {
      "delete": {
        "tags": [
          "后台管理"
        ],
        "summary": "批量软删除用户",
        "description": "将明确选中的用户账户移入回收站并停用，记录可选删除原因并撤销其全部会话。",
        "operationId": "delete_users_bulk_api_v1_admin_users_batch_delete",
        "parameters": [
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserBulkDeleteIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_BatchActionResult_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/restore/batch": {
      "post": {
        "tags": [
          "后台管理"
        ],
        "summary": "批量恢复回收站用户",
        "description": "恢复回收站中的用户，恢复后账户保持停用。",
        "operationId": "restore_users_bulk_api_v1_admin_users_restore_batch_post",
        "parameters": [
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRestoreBatchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_BatchActionResult_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/restore": {
      "post": {
        "tags": [
          "后台管理"
        ],
        "summary": "恢复回收站用户",
        "description": "恢复回收站中的用户，恢复后账户保持停用。",
        "operationId": "restore_user_api_v1_admin_users__user_id__restore_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminUserRead_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail",
            "description": "请求参数校验错误详情列表"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResponseModel_AdminUserRead_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "面向调用方的中文结果消息"
          },
          "data": {
            "$ref": "#/components/schemas/AdminUserRead",
            "description": "响应业务数据"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "用于定位本次请求的唯一标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[AdminUserRead]"
      },
...
      "ResponseModel_PageResult_AdminUserRead__": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "面向调用方的中文结果消息"
          },
          "data": {
            "$ref": "#/components/schemas/PageResult_AdminUserRead_",
            "description": "响应业务数据"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "用于定位本次请求的唯一标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[PageResult[AdminUserRead]]"
      },
...
    }
  },
...
}